Refactoring, autodeletion of temp filesystem, fix for issue 13827
svn path=/trunk/; revision=41919


Former-commit-id: 58b4fecd39fdbf336d37ee21839bb18cdf3afa6b

+++ v3/common/glassfish-api/src/main/java/org/glassfish/api/embedded/EmbeddedFileSystem.java
@@ -197,7 +197,7 @@ public class EmbeddedFileSystem implements PreDestroy {
 
     public void preDestroy() {
         Logger.getAnonymousLogger().finer("delete " + instanceRoot + " = " + autoDelete);
-        if (autoDelete) {
+        if (autoDelete && instanceRoot != null) {
             // recursively delete instanceRoot directory
             Logger.getAnonymousLogger().finer("Deleting recursively" + instanceRoot);
             deleteAll(instanceRoot);
